requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
filelock==3.3.2
fonttools==4.29.0
grpcio==1.41.1
gym==0.10.5
idna==3.3
iniconfig==1.1.1
kiwisolver==1.3.2
matplotlib==3.5.1
msgpack==1.0.2
neat-python==0.92
numpy==1.21.3
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyglet==1.5.21
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
ray==1.7.0
redis==3.5.3
requests==2.26.0
scipy==1.7.1
six==1.16.0
toml==0.10.2
torch==1.10.0
typing-extensions==3.10.0.2
urllib3==1.26.7
